rename : ReservationRequest -> ReservationCheck
